requirements.txt
Django
djangorestframework
psycopg2
BeautifulSoup4
requests
lxml
coverage
python-dateutil
django-cors-headers